refactor(nav): place all navigation menu in one dir
