--- src/layouts/components/UserMenu/styles.css
@@ -1,4 +1,0 @@
-.UserButton {
-    margin-left: 8px;
-    border-radius: 50%;
-}
